Move WellKnownTypes inside IXamlTypeSystem
(cherry picked from commit ec7b5f35f07ea57983ad8efc0e9df53a0f051f98)
