+++ src/XamlX.IL.Cecil/CecilType.cs
@@ -35,7 +35,7 @@ namespace XamlX.TypeSystem
                 Reference = reference;
                 Definition = definition;
                 if (reference.IsArray)
-                    Definition = ((CecilType)TypeSystem.GetType("System.Array")).Definition;
+                    Definition = ((CecilType)TypeSystem.WellKnownTypes.Array).Definition;
 
                 TypeResolveContext = parentTypeResolveContext.Nested(reference);
             }
